add: missed functions during reset
